# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Central Studio & Stage Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      parameters:
        - { name: classId, in: query, required: false, schema: { type: integer } }
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /schedules/{id}:
    get:
      operationId: getSchedule
      tags: [schedules]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSchedule
      tags: [schedules]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduleBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /price-packages:
    get:
      operationId: listPricePackages
      tags: [pricePackages]
      responses:
        "200":
          description: List of price packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PricePackage"
    post:
      operationId: createPricePackage
      tags: [pricePackages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePricePackageBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricePackage"
  /price-packages/{id}:
    get:
      operationId: getPricePackage
      tags: [pricePackages]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: PricePackage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricePackage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePricePackage
      tags: [pricePackages]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePricePackageBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricePackage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePricePackage
      tags: [pricePackages]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateStudent
      tags: [students]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStudent
      tags: [students]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Schedule:
      type: object
      required: [id, classId, dayOfWeek, startTime, endTime, isRecurring, createdAt]
      properties:
        id: { type: integer }
        classId: { type: integer }
        dayOfWeek: { type: integer }
        startTime: { type: string }
        endTime: { type: string }
        location: { type: ["string", "null"] }
        isRecurring: { type: boolean }
        effectiveFrom: { type: ["string", "null"] }
        effectiveUntil: { type: ["string", "null"] }
        createdAt: { type: string }
    CreateScheduleBody:
      type: object
      required: [classId, dayOfWeek, startTime, endTime]
      properties:
        classId: { type: integer }
        dayOfWeek: { type: integer }
        startTime: { type: string }
        endTime: { type: string }
        location: { type: ["string", "null"] }
        isRecurring: { type: boolean }
        effectiveFrom: { type: ["string", "null"] }
        effectiveUntil: { type: ["string", "null"] }
    UpdateScheduleBody:
      type: object
      properties:
        classId: { type: integer }
        dayOfWeek: { type: integer }
        startTime: { type: string }
        endTime: { type: string }
        location: { type: ["string", "null"] }
        isRecurring: { type: boolean }
        effectiveFrom: { type: ["string", "null"] }
        effectiveUntil: { type: ["string", "null"] }
    PricePackage:
      type: object
      required: [id, name, type, priceEgp, isActive, isFeatured, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        priceEgp: { type: number }
        sessions: { type: ["integer", "null"] }
        description: { type: ["string", "null"] }
        isActive: { type: boolean }
        isFeatured: { type: boolean }
        createdAt: { type: string }
    CreatePricePackageBody:
      type: object
      required: [name, type, priceEgp]
      properties:
        name: { type: string }
        type: { type: string }
        priceEgp: { type: number }
        sessions: { type: ["integer", "null"] }
        description: { type: ["string", "null"] }
        isActive: { type: boolean }
        isFeatured: { type: boolean }
    UpdatePricePackageBody:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        priceEgp: { type: number }
        sessions: { type: ["integer", "null"] }
        description: { type: ["string", "null"] }
        isActive: { type: boolean }
        isFeatured: { type: boolean }
...
    Student:
      type: object
      required: [id, name, email, totalBookings, joinedAt, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        totalBookings: { type: integer }
        joinedAt: { type: string }
        createdAt: { type: string }
    CreateStudentBody:
      type: object
      required: [name, email]
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    UpdateStudentBody:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        totalBookings: { type: integer }
...